requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
httpx==0.24.1
requests==2.31.0
slack-sdk==3.21.3
pydantic==2.3.0
redis==4.6.0
PyJWT==2.8.0
sqlalchemy==2.0.20
python-json-logger==2.0.7
psutil==5.9.6
bleach==6.1.0
